about section updated
