+++ components/About/AboutMe.tsx
@@ -1,39 +1,17 @@
 "use client";
 
-import { personalInfo } from "@/data/portfolio";
+import {
+  personalInfo,
+  fadeIn,
+  slideInLeft,
+  slideInRight,
+} from "@/data/portfolio";
 import { MapPin, Briefcase, Target, Heart } from "lucide-react";
 import { motion } from "framer-motion";
 
 interface AboutMeProps {
   t: any;
 }
-
-// Animation variants for Framer Motion
-const fadeIn = {
-  hidden: { opacity: 0 },
-  visible: {
-    opacity: 1,
-    transition: { duration: 0.6, ease: "easeInOut" },
-  },
-};
-
-const slideInLeft = {
-  hidden: { opacity: 0, x: -50 },
-  visible: {
-    opacity: 1,
-    x: 0,
-    transition: { duration: 0.7, ease: "easeOut" },
-  },
-};
-
-const slideInRight = {
-  hidden: { opacity: 0, x: 50 },
-  visible: {
-    opacity: 1,
-    x: 0,
-    transition: { duration: 0.7, ease: "easeOut" },
-  },
-};
 
 export const AboutMe = ({ t }: AboutMeProps) => {
   const quickFactsData = [
